+++ ql/cashflows/zeroinflationcashflow.cpp
@@ -17,7 +17,7 @@
  FOR A PARTICULAR PURPOSE.  See the license for more details.
 */
 
-#include <ql/cashflows/inflationcashflow.hpp>
+#include <ql/cashflows/zeroinflationcashflow.hpp>
 #include <ql/indexes/inflationindex.hpp>
 #include <ql/termstructures/inflationtermstructure.hpp>
 #include <ql/time/calendars/nullcalendar.hpp>
